refactor: upgrade Fresh v1.3 and use plugin
