+++ main.ts
@@ -9,4 +9,6 @@ import "$std/dotenv/load.ts";
 import { start } from "$fresh/server.ts";
 import manifest from "./fresh.gen.ts";
 
-await start(manifest);
+import { plugin } from "./utils/kv_oauth.ts";
+
+await start(manifest, { plugins: [plugin()] });
