--- routes/signin.ts
@@ -1,9 +1,0 @@
-import type { Handlers } from "$fresh/server.ts";
-import { signIn } from "kv_oauth";
-import { oauth2Client } from "../utils/oauth2_client.ts";
-
-export const handler: Handlers = {
-  async GET(req) {
-    return await signIn(req, oauth2Client);
-  },
-};
