--- utils/oauth2_client.ts
@@ -1,3 +1,0 @@
-import { createGitHubOAuth2Client } from "kv_oauth";
-
-export const oauth2Client = createGitHubOAuth2Client();
